Current efforts to get things working
Right now, pretty much everything is working except for the ServiceMonitor configuration.
For some reason the same way I did it in telemetry-framework-apb doesn't seem to work here.

I'm wondering if it has something to do with how I was doing the setup with Prometheus in
the other playbook, issue with k8s module not working the same as oc_obj, or something else
I haven't been able to track down yet.: